Refactor Bonsai.ONI/clroni interaction
- Previously, each node that required interaction with an ONI context
had the power to create the context
- When other nodes were added, they would see that the context was
opened and receive a reference to it
- The references were counted until all nodes did not need them
- This creates big issues, e.g. if a node is repeated subscribed and
unscibscribed from as it causes a reinitialization of the context every
time, among other things
- To deal with this, I followed the changes made to the Arduino library
in Bonsai, to some extent
- Now, there is a CreateXXContent node (only CreateXillyContext now, but
will be expanded as library of drivers expands) that explicity creates
an ONI context. All ONI device nodes must be connected, physically to
this context via an Observable Sequence. This makes the chain of effect
much more transparant, espeically when dealing with multiple contexts
and makes the devices much less stateful
- Further, the devices themselves have been largely subsummed by
generic functionality in ONIDevice & ObservableFrame. This makes devices
much shorter an more maintainable
- The Electrical stimulator has not been fully tested and the
OpticalStimulator has been removed for now
- The prototype project has been removed from the solution until the
main refactor is complete